+++ app/src/main/java/com/sstechcanada/todo/activities/auth/ProfileActivity.kt
@@ -22,19 +22,16 @@ import androidx.appcompat.widget.AppCompatTextView
 import androidx.preference.PreferenceManager
 import com.bumptech.glide.Glide
 import com.google.android.gms.ads.*
-import com.google.android.gms.ads.rewarded.RewardedAd
-import com.google.android.gms.ads.rewarded.RewardedAdLoadCallback
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.google.android.material.button.MaterialButton
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.SetOptions
 import com.sstechcanada.todo.BuildConfig
 import com.sstechcanada.todo.R
 import com.sstechcanada.todo.activities.AboutActivity
 import com.sstechcanada.todo.activities.MasterTodoListActivity
+import com.sstechcanada.todo.activities.RemoveAdsActivity
 import com.sstechcanada.todo.activities.SplashActivity
 import com.sstechcanada.todo.utils.SaveSharedPreference
 import es.dmoral.toasty.Toasty
@@ -44,7 +41,6 @@ class ProfileActivity : AppCompatActivity() {
 
     private var mAuth: FirebaseAuth? = null
     private var mGoogleSignInClient: GoogleSignInClient? = null
-    private var mRewardedAd: RewardedAd? = null
     private val TAG = "ProfileActivity"
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -81,7 +77,9 @@ class ProfileActivity : AppCompatActivity() {
         cardRate.setOnClickListener { openGooglePlayForRating() }
         cardDelete.setOnClickListener { deleteWarningDialog() }
         cardSignOut.setOnClickListener { showSignOutDialog() }
-        cardRemoveAds.setOnClickListener { showRewardedVideo() }
+        cardRemoveAds.setOnClickListener {
+            startActivity(Intent(this@ProfileActivity, RemoveAdsActivity::class.java))
+        }
     }
 
     private fun setupFirebaseLogin() {
@@ -96,12 +94,9 @@ class ProfileActivity : AppCompatActivity() {
     private fun setupAds() {
         if (SaveSharedPreference.getAdsEnabled(this)) {
             adView.loadAd(AdRequest.Builder().build())
-            loadRewardedAd()
             adView.visibility = View.VISIBLE
-            cardRemoveAds.visibility = View.VISIBLE
         } else {
             adView.visibility = View.GONE
-            cardRemoveAds.visibility = View.GONE
         }
 
     }
@@ -219,71 +214,6 @@ class ProfileActivity : AppCompatActivity() {
         dialog.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
         dialog.window?.attributes?.windowAnimations = R.style.DialogAnimation
         dialog.window?.setGravity(Gravity.BOTTOM)
-    }
-
-    private fun loadRewardedAd() {
-        val adRequest = AdRequest.Builder().build()
-        RewardedAd.load(
-            this,
-            getString(R.string.rewarded_ad_unit_id),
-            adRequest,
-            object : RewardedAdLoadCallback() {
-                override fun onAdFailedToLoad(adError: LoadAdError) {
-                    Log.d(TAG, adError.toString())
-                    mRewardedAd = null
-                }
-
-                override fun onAdLoaded(rewardedAd: RewardedAd) {
-                    Log.d(TAG, "Ad was loaded.")
-                    mRewardedAd = rewardedAd
-                }
-            })
-    }
-
-    private fun showRewardedVideo() {
-        if (mRewardedAd != null) {
-            mRewardedAd?.fullScreenContentCallback =
-                object : FullScreenContentCallback() {
-                    override fun onAdDismissedFullScreenContent() {
-                        Log.d(TAG, "Ad was dismissed.")
-                        mRewardedAd = null
-                        loadRewardedAd()
-                        val moveToMasterTodoListActivity = Intent(this@ProfileActivity, MasterTodoListActivity::class.java)
-                        startActivity(moveToMasterTodoListActivity)
-                        finish()
-                    }
-
-                    override fun onAdFailedToShowFullScreenContent(adError: AdError) {
-                        Log.d(TAG, "Ad failed to show.")
-                        mRewardedAd = null
-                    }
-
-                    override fun onAdShowedFullScreenContent() {
-                        Log.d(TAG, "Ad showed fullscreen content.")
-                    }
-                }
-
-            mRewardedAd?.show(this) {
-                val data: MutableMap<String, String?> = HashMap()
-                data["adsPausedTimestamp"] = getTimeStampOfNextWeek()
-                data["purchase_code"] = "3"
-                mAuth?.currentUser?.uid?.let {
-                    FirebaseFirestore.getInstance().collection("Users").document(it).set(
-                        data,
-                        SetOptions.merge()
-                    ).addOnSuccessListener {
-                        Toasty.success(
-                            this,
-                            "Ads are paused for a week",
-                            Toast.LENGTH_LONG
-                        ).show()
-                    }
-                }
-                Log.d("TAG", "User earned the reward.")
-            }
-        } else {
-            Toasty.info(this, "No Ads available, try again later.", Toast.LENGTH_SHORT).show()
-        }
     }
 
     private fun getTimeStampOfNextWeek(): String {
