+++ app/src/main/java/com/sstechcanada/todo/utils/RemoveAdsUtils.kt
@@ -1,0 +1,10 @@
+package com.sstechcanada.todo.utils
+
+object RemoveAdsUtils {
+
+    fun getTimeStampOfNextWeek(): String {
+        val now = System.currentTimeMillis() / 1000
+        val afterAWeek = now + 60//604800
+        return afterAWeek.toString()
+    }
+}
